+++ example/lib/vote.dart
@@ -7,6 +7,7 @@ Future<void> vote() async {
   GatewayPool gw;
   EntityMetadata entityMeta;
   ProcessMetadata processMeta;
+  ProcessData processData;
   EnvelopePackage pollVoteEnvelope;
   String merkleProof;
   int blockHeight, censusSize, envelopeHeight;
@@ -42,7 +43,8 @@ Future<void> vote() async {
       print("There are no active processes");
       return;
     }
-    processMeta = await getProcessMetadata(pid, gw);
+    processData = await getProcess(pid, gw);
+    processMeta = await getProcessMetadata(pid, gw, data: processData);
     processMeta.meta["id"] = pid;
     print("Process ID: $pid");
   } catch (err) {
@@ -60,7 +62,7 @@ Future<void> vote() async {
 
     // Census size
     print("\nQuerying for the Census size");
-    censusSize = await getCensusSize(processMeta.census.merkleRoot, gw);
+    censusSize = await getCensusSize(processData.getCensusRoot, gw);
     // censusSize = await getCensusSize(censusMerkleRoot, gw);
     if (!(censusSize is int)) throw Exception("The census size is not valid");
     print("Census size: $censusSize");
@@ -74,17 +76,17 @@ Future<void> vote() async {
 
     // Remaining seconds
     print("\nEstimating");
-    dateAtBlock = await estimateDateAtBlock(processMeta.startBlock, gw);
+    dateAtBlock = await estimateDateAtBlock(processData.getStartBlock, gw);
     print("Process start block: $dateAtBlock");
     dateAtBlock = await estimateDateAtBlock(
-        processMeta.startBlock + processMeta.blockCount, gw);
+        processData.getStartBlock + processData.getBlockCount, gw);
     print("Process end block: $dateAtBlock");
 
     // Merkle Proof
     print("\nRequesting Merkle Proof");
     final isDigested = true;
     merkleProof = await generateProof(
-        processMeta.census.merkleRoot, pubKeyClaim, isDigested, gw);
+        processData.getCensusRoot, pubKeyClaim, isDigested, gw);
     // merkleProof = await generateProof(censusMerkleRoot, pubKeyClaim, gw);
     if (!(merkleProof is String))
       throw Exception("The Merkle Proof is not valid");